test(http-linter): move integration tests to own file
