+++ packages/http-linter/test/analytics-integration.test.ts
@@ -386,77 +386,4 @@ describe('http-linter analytics integration', { timeout: 30000 }, () => {
     expect(result.valid).toBeFalsy();
     expect(result.reports).toHaveLength(50);
   });
-
-  it('should provide options to rules', async () => {
-    await thymian
-      .register(httpLinterPlugin, {
-        ruleSets: [
-          join(
-            import.meta.dirname,
-            'fixtures/rules/rule-with-options.rule.mjs',
-          ),
-        ],
-        type: ['static'],
-        rules: {
-          'rule-with-options': {
-            options: {
-              foo: 'lol',
-            },
-          },
-        },
-      })
-      .ready();
-
-    const format = createThymianFormatWithTransaction(
-      createHttpRequest({
-        headers: {
-          lol: {
-            required: false,
-            schema: {
-              type: 'string',
-            },
-            style: DEFAULT_HEADER_SERIALIZATION_STYLE,
-          },
-        },
-      }),
-      createHttpResponse(),
-    );
-
-    const result = await thymian.emitter.emitAction(
-      'http-linter.lint-static',
-      { format: format.export() },
-      { strategy: 'first' },
-    );
-
-    expect(result.valid).toBeFalsy();
-  });
-
-  it.each([
-    ['off', 0],
-    ['warn', 2],
-    ['error', 1],
-    ['hint', 2],
-  ])(
-    'should load rules respecting given severity %s and return %d rules',
-    async (severity, length) => {
-      await thymian
-        .register(httpLinterPlugin, {
-          ruleSets: [
-            join(import.meta.dirname, 'fixtures/rules/a.rule.mjs'),
-            join(import.meta.dirname, 'fixtures/rules/b.rule.mjs'),
-          ],
-          severity: severity as RuleSeverity,
-          type: ['test'],
-        })
-        .ready();
-
-      const result = await thymian.emitter.emitAction(
-        'http-linter.rules',
-        undefined,
-        { strategy: 'first' },
-      );
-
-      expect(result).toHaveLength(length);
-    },
-  );
 });
